Simplify Splash Screen Launch for integration call
