+++ src/DynamoCoreWpf/DynamoCoreWpf.csproj
@@ -1398,6 +1398,7 @@
     </Page>
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\DynamoApplications\DynamoApplications.csproj" />
     <ProjectReference Include="..\DynamoCore\DynamoCore.csproj">
       <Project>{7858fa8c-475f-4b8e-b468-1f8200778cf8}</Project>
       <Name>DynamoCore</Name>
