+++ src/DynamoCoreWpf/Views/GuidedTour/ExitGuideWindow.xaml.cs
@@ -1,19 +1,6 @@
-﻿using Dynamo.Wpf.ViewModels.GuidedTour;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+using Dynamo.Wpf.ViewModels.GuidedTour;
 
 namespace Dynamo.Wpf.Views.GuidedTour
 {
